refact: utils 책임 분리
