+++ src/features/calendar/utils/index.ts
@@ -1,0 +1,15 @@
+/**
+ * 캘린더 관련 유틸리티 함수 내보내기
+ */
+
+// 세션 관련 유틸
+export * from "./sessionUtils";
+
+// 날짜 관련 유틸
+export * from "./dateUtils";
+
+// 포맷팅 관련 유틸
+export * from "./formatUtils";
+
+// 서비스로 이동된 SessionStorageService - 이전 코드와의 호환성을 위해 재내보내기
+export { sessionStorageService } from "../services/storage/SessionStorageService";
